+++ src/core/ProofSet.cpp
@@ -46,13 +46,13 @@ std::vector<uint64_t>& Words::data() noexcept {
     return data_;
 }
 
-Words Words::fromUint64(const std::vector<uint64_t>& host, uint64_t exponent) {
+Words Words::fromUint64(const std::vector<uint64_t>& host, uint32_t exponent) {
     (void)exponent;
     return Words(host);
 }
 
 // ProofSet
-ProofSet::ProofSet(uint64_t exponent, uint64_t proofLevel)
+ProofSet::ProofSet(uint32_t exponent, uint32_t proofLevel)
   : E{exponent}, power{proofLevel} {
   if(exponent%2!=0){
       assert(E & 1); // E is supposed to be prime
@@ -61,14 +61,14 @@ ProofSet::ProofSet(uint64_t exponent, uint64_t proofLevel)
     std::filesystem::create_directories(proofPath(E));
 
     // Calculate checkpoint points using binary tree structure
-    std::vector<uint64_t> spans;
-    for (uint64_t span = (E + 1) / 2; spans.size() < power; span = (span + 1) / 2) { 
+    std::vector<uint32_t> spans;
+    for (uint32_t span = (E + 1) / 2; spans.size() < power; span = (span + 1) / 2) { 
       spans.push_back(span); 
     }
 
     points.push_back(0);
-    for (uint64_t p = 0, span = (E + 1) / 2; p < power; ++p, span = (span + 1) / 2) {
-      for (uint64_t i = 0, end = static_cast<uint64_t>(points.size()); i < end; ++i) {
+    for (uint32_t p = 0, span = (E + 1) / 2; p < power; ++p, span = (span + 1) / 2) {
+      for (uint32_t i = 0, end = static_cast<uint32_t>(points.size()); i < end; ++i) {
         points.push_back(points[i] + span);
       }
     }
@@ -82,20 +82,20 @@ ProofSet::ProofSet(uint64_t exponent, uint64_t proofLevel)
     assert(points.size() == (1u << power));
     assert(points.back() == E);
 
-    points.push_back(uint64_t(-1)); // guard element
+    points.push_back(uint32_t(-1)); // guard element
 
     // Verify all points are valid
-    for (uint64_t p : points) {
+    for (uint32_t p : points) {
       assert(p > E || isInPoints(E, power, p));
     }
   }
 }
 
-bool ProofSet::shouldCheckpoint(uint64_t iter) const {
+bool ProofSet::shouldCheckpoint(uint32_t iter) const {
   return isInPoints(E, power, iter);
 }
 
-void ProofSet::save(uint64_t iter, const std::vector<uint64_t>& words) {
+void ProofSet::save(uint32_t iter, const std::vector<uint32_t>& words) {
   if (!shouldCheckpoint(iter)) {
     return;
   }
@@ -110,21 +110,21 @@ void ProofSet::save(uint64_t iter, const std::vector<uint64_t>& words) {
   }
   
   // Write CRC32 first, then the data
-  uint64_t crc = computeCRC32(words.data(), words.size() * sizeof(uint64_t));
+  uint32_t crc = computeCRC32(words.data(), words.size() * sizeof(uint32_t));
   file.write(reinterpret_cast<const char*>(&crc), sizeof(crc));
   file.write(reinterpret_cast<const char*>(words.data()), 
-             words.size() * sizeof(uint64_t));
+             words.size() * sizeof(uint32_t));
   
   if (!file.good()) {
     throw std::runtime_error("Error writing proof checkpoint file: " + filePath.string());
   }
 }
 
-Words ProofSet::fromUint64(const std::vector<uint64_t>& host, uint64_t exponent) {
+Words ProofSet::fromUint64(const std::vector<uint64_t>& host, uint32_t exponent) {
     return Words::fromUint64(host, exponent);
 }
 
-uint64_t ProofSet::bestPower(uint64_t E) {
+uint32_t ProofSet::bestPower(uint32_t E) {
   // Best proof powers assuming no disk space concern.
   // We increment power by 1 for each fourfold increase of the exponent.
   // The values below produce power=10 at wavefront, and power=11 at 100Mdigits:
@@ -135,13 +135,13 @@ uint64_t ProofSet::bestPower(uint64_t E) {
   int32_t power = 10 + static_cast<int32_t>(std::floor(std::log2(E / 60e6) / 2));
   power = std::max(power, 2);
   power = std::min(power, 12);
-  return static_cast<uint64_t>(power);
+  return static_cast<uint32_t>(power);
 }
 
-bool ProofSet::isInPoints(uint64_t E, uint64_t power, uint64_t k) {
+bool ProofSet::isInPoints(uint32_t E, uint32_t power, uint32_t k) {
   if (k == E) { return true; } // special-case E
-  uint64_t start = 0;
-  for (uint64_t p = 0, span = (E + 1) / 2; p < power; ++p, span = (span + 1) / 2) {
+  uint32_t start = 0;
+  for (uint32_t p = 0, span = (E + 1) / 2; p < power; ++p, span = (span + 1) / 2) {
     assert(k >= start);
     if (k > start + span) {
       start += span;
@@ -152,13 +152,13 @@ bool ProofSet::isInPoints(uint64_t E, uint64_t power, uint64_t k) {
   return false;
 }
 
-std::filesystem::path ProofSet::proofPath(uint64_t E) {
+std::filesystem::path ProofSet::proofPath(uint32_t E) {
   return std::filesystem::path(std::to_string(E)) / "proof";
 }
 
-bool ProofSet::isValidTo(uint64_t limitK) const {
+bool ProofSet::isValidTo(uint32_t limitK) const {
   // Check if we have all required checkpoint files up to limitK
-  for (uint64_t point : points) {
+  for (uint32_t point : points) {
     if (point > limitK) break;
     if (point < E && !fileExists(point)) {
       return false;
@@ -167,12 +167,12 @@ bool ProofSet::isValidTo(uint64_t limitK) const {
   return true;
 }
 
-bool ProofSet::fileExists(uint64_t k) const {
+bool ProofSet::fileExists(uint32_t k) const {
   auto filePath = proofPath(E) / std::to_string(k);
   return std::filesystem::exists(filePath);
 }
 
-std::vector<uint64_t> ProofSet::load(uint64_t iter) const {
+std::vector<uint32_t> ProofSet::load(uint32_t iter) const {
   if (!shouldCheckpoint(iter)) {
     throw std::runtime_error("Attempt to load non-checkpoint iteration: " + std::to_string(iter));
   }
@@ -184,24 +184,24 @@ std::vector<uint64_t> ProofSet::load(uint64_t iter) const {
   }
 
   // Read CRC32 first
-  uint64_t crc;
+  uint32_t crc;
   file.read(reinterpret_cast<char*>(&crc), sizeof(crc));
   if (!file.good()) {
     throw std::runtime_error("Error reading CRC32 from proof checkpoint file: " + filePath.string());
   }
 
   // Calculate expected file size in 32-bit words: (E + 31) / 32
-  uint64_t expectedWords = (E + 31) / 32;
+  uint32_t expectedWords = (E + 31) / 32;
   
   // Read the 32-bit words data
-  std::vector<uint64_t> words(expectedWords);
-  file.read(reinterpret_cast<char*>(words.data()), expectedWords * sizeof(uint64_t));
+  std::vector<uint32_t> words(expectedWords);
+  file.read(reinterpret_cast<char*>(words.data()), expectedWords * sizeof(uint32_t));
   if (!file.good()) {
     throw std::runtime_error("Error reading data from proof checkpoint file: " + filePath.string());
   }
 
   // Verify CRC32
-  uint64_t computedCrc = computeCRC32(words.data(), words.size() * sizeof(uint64_t));
+  uint32_t computedCrc = computeCRC32(words.data(), words.size() * sizeof(uint32_t));
   if (crc != computedCrc) {
     throw std::runtime_error("CRC32 mismatch in proof checkpoint file: " + filePath.string());
   }
@@ -213,7 +213,7 @@ Proof ProofSet::computeProof() const {
   // Start timing proof generation
   util::Timer timer;
 
-  std::vector<std::vector<uint64_t>> middles;
+  std::vector<std::vector<uint32_t>> middles;
   std::vector<uint64_t> hashes;
 
   // Initial hash of the final residue B
@@ -221,32 +221,32 @@ Proof ProofSet::computeProof() const {
   auto hash = Proof::hashWords(E, B);
 
   // Pre-allocate maximum needed buffer pool (power levels use 2^p buffers max)
-  uint64_t maxBuffers = (1u << power);
+  uint32_t maxBuffers = (1u << power);
   std::vector<mpz_class> bufferPool(maxBuffers);
 
   // Main computation loop
-  for (uint64_t p = 0; p < power; ++p) {
+  for (uint32_t p = 0; p < power; ++p) {
     assert(p == hashes.size());
     
-    uint64_t s = (1u << (power - p - 1)); // Step size for this level
-    uint64_t levelBuffers = (1u << p); // Number of buffers needed for this level
-    uint64_t bufIndex = 0;
+    uint32_t s = (1u << (power - p - 1)); // Step size for this level
+    uint32_t levelBuffers = (1u << p); // Number of buffers needed for this level
+    uint32_t bufIndex = 0;
     
     // Clear buffers that will be used for this level
-    for (uint64_t i = 0; i < levelBuffers; ++i) {
+    for (uint32_t i = 0; i < levelBuffers; ++i) {
       bufferPool[i] = 0;
     }
     
     // Load residues and apply binary tree algorithm
-    for (uint64_t i = 0; i < levelBuffers; ++i) {
+    for (uint32_t i = 0; i < levelBuffers; ++i) {
       // PRPLL's formula: load checkpoint at points[s * (i * 2 + 1) - 1]
-      uint64_t checkpointIndex = s * (i * 2 + 1) - 1;
+      uint32_t checkpointIndex = s * (i * 2 + 1) - 1;
       
       if (checkpointIndex >= points.size()) {
         continue;
       }
       
-      uint64_t iteration = points[checkpointIndex];
+      uint32_t iteration = points[checkpointIndex];
       
       if (iteration > E || !shouldCheckpoint(iteration)) {
         continue;
@@ -257,7 +257,7 @@ Proof ProofSet::computeProof() const {
       bufIndex++;
       
       // Apply hashes from previous levels
-      for (uint64_t k = 0; i & (1u << k); ++k) {
+      for (uint32_t k = 0; i & (1u << k); ++k) {
         assert(k <= p - 1);
         if (bufIndex < 2) {
           std::cerr << "Error: need at least 2 buffers for expMul, have " << bufIndex << std::endl;
@@ -309,16 +309,16 @@ Proof ProofSet::computeProof() const {
   return Proof{E, std::move(B), std::move(middles)};
 }
 
-mpz_class ProofSet::convertToGMP(const std::vector<uint64_t>& words) const {
+mpz_class ProofSet::convertToGMP(const std::vector<uint32_t>& words) const {
   mpz_class result;
   // Use GMP's optimized mpz_import function
-  mpz_import(result.get_mpz_t(), words.size(), -1 /*order: LSWord first*/, sizeof(uint64_t), 0 /*endian: native*/, 0 /*nails*/, words.data());
+  mpz_import(result.get_mpz_t(), words.size(), -1 /*order: LSWord first*/, sizeof(uint32_t), 0 /*endian: native*/, 0 /*nails*/, words.data());
   return result;
 }
 
 // Optimized modular reduction for Mersenne numbers: x mod (2^E - 1)
 // Uses the identity: X mod (2^E - 1) ≡ (Xlo + Xhi) mod (2^E - 1)
-mpz_class ProofSet::mersenneReduce(const mpz_class& x, uint64_t E) const {
+mpz_class ProofSet::mersenneReduce(const mpz_class& x, uint32_t E) const {
   // For small numbers, use regular mod
   if (mpz_sizeinbase(x.get_mpz_t(), 2) <= E + 1) {
     return x;
@@ -349,7 +349,7 @@ mpz_class ProofSet::mersenneReduce(const mpz_class& x, uint64_t E) const {
 
 // Optimized modular exponentiation for Mersenne numbers: base^exp mod (2^E - 1)
 // Uses fast Mersenne reduction at each step instead of general division
-mpz_class ProofSet::mersennePowMod(const mpz_class& base, uint64_t exp, uint64_t E) const {
+mpz_class ProofSet::mersennePowMod(const mpz_class& base, uint64_t exp, uint32_t E) const {
   if (exp == 0) {
     return mpz_class(1);
   }
@@ -383,20 +383,20 @@ mpz_class ProofSet::mersennePowMod(const mpz_class& base, uint64_t exp, uint64_t
   return result;
 }
 
-std::vector<uint64_t> ProofSet::convertFromGMP(const mpz_class& gmp_val) const {
+std::vector<uint32_t> ProofSet::convertFromGMP(const mpz_class& gmp_val) const {
   size_t wordCount = (E + 31) / 32;
-  std::vector<uint64_t> data(wordCount, 0);
+  std::vector<uint32_t> data(wordCount, 0);
   
   // Use GMP's optimized mpz_export function
   size_t actualWords = 0;
-  mpz_export(data.data(), &actualWords, -1 /*order: LSWord first*/, sizeof(uint64_t), 0 /*endian: native*/, 0 /*nails*/, gmp_val.get_mpz_t());
+  mpz_export(data.data(), &actualWords, -1 /*order: LSWord first*/, sizeof(uint32_t), 0 /*endian: native*/, 0 /*nails*/, gmp_val.get_mpz_t());
   
   // Note: actualWords may be less than wordCount if the number has leading zeros
   // The vector is already zero-initialized, so this is correct
   return data;
 }
 
-double ProofSet::diskUsageGB(uint64_t E, uint64_t power) {
+double ProofSet::diskUsageGB(uint32_t E, uint32_t power) {
   // Calculate disk usage in GB for proof files
   // Formula from PRPLL: ldexp(E, -33 + int(power)) * 1.05
   if (power == 0) return 0.0;
